Fix bug with Util.findBoundaries()
Function assumed both boundaries existed in the source string.  If they
didn't, you'd end up getting an unexpected result or an IndexOutOfBounds
exception.  Function would also throw an NPE if str was null.

This commit address all of the issues above, throwing a descriptive
IllegalArgumentException when the arguments given don't allow for an
expected result to be returned; also avoids throwing less than
desireable IndexOutOfBounds or NPE exceptions as well.